EOS-17964: Implement event manager interface to subscribe and unsubscribe HA events
- Update the code to store new structured key-value pair to consul
- Rebase
